+++ front/app/components/admin/ContentBuilder/Widgets/ThreeColumn/index.tsx
@@ -8,11 +8,12 @@ import {
 } from '@citizenlab/cl2-component-library';
 
 // craft
-import { Element } from '@craftjs/core';
+import { Element, ROOT_NODE, useNode } from '@craftjs/core';
 import Container from '../Container';
 
 // i18n
 import messages from './messages';
+import { DEFAULT_PADDING } from '../../constants';
 
 export const ThreeColumnWrapper = ({
   children,
@@ -37,8 +38,17 @@ export const ThreeColumnWrapper = ({
 };
 
 const ThreeColumn = () => {
+  const isSmallerThanTablet = useBreakpoint('tablet');
+  const { parent } = useNode((node) => ({
+    parent: node.data.parent,
+  }));
+
   return (
-    <ThreeColumnWrapper>
+    <ThreeColumnWrapper
+      maxWidth="1150px"
+      margin="0 auto"
+      px={isSmallerThanTablet && parent === ROOT_NODE ? DEFAULT_PADDING : '0px'}
+    >
       <Box flex="1">
         <Element id="column1" is={Container} canvas />
       </Box>
